Setup repository and add demo Cypress test
Setting up the Cypress folder structure and other repository dependencies such as TypeScript and ESLint. This adds a working Cypress functional test to run on OpenSearch Dashboards, and updates the .gitignore to include local Cypress files. As well, this adds instructions to the README on how to setup and run the Cypress tests.

Signed-off-by: Aviv Benchorin <benchori@amazon.com>