Add dashboard module to demo functional test (#7)
Adding new dashboard test suite to the demo functional test, as well as updating the method of verifying the OpenSearch Dashboards home page has loaded properly.

Adds known flakiness through the testing of the EuiComboBox element, which introduces inconsistent behavior with the ComboBox “isinvalid” state when trying to click on an option after entering text in the search box. This bug is further discussed in #8.

Signed-off-by: Aviv Benchorin <benchori@amazon.com>