Add dashboard_filtering_spec.js functional test
Adds a new Cypress test file dashboard_filtering_spec.js, which tests whether dashboard visualizations properly update when applying and disabling filters. This functional test is a Cypress adaptation of the dashboard_filtering.js test file from OpenSearch Dashboards, which originally used the Selenium framework. The mapping and index archives used in the OpenSearch Dashboards test have been added to the fixtures directory.

The test file is incomplete, and lacks the functionality to import the indices and data used in the test to OpenSearch. The test also contains known inconsistent behavior with combo boxes (described in #8).

Signed-off-by: Aviv Benchorin <benchori@amazon.com>