Add dashboard_filtering_spec.js functional test (#14)
Adds a new Cypress test file dashboard_filtering_spec.js, which tests whether dashboard visualizations properly update when applying and disabling filters. This functional test is a Cypress adaptation of the dashboard_filtering.js test file from OpenSearch Dashboards, which originally used the Selenium framework. The mapping and index archives used in the OpenSearch Dashboards test have been added to the fixtures directory.

The test file is incomplete, and lacks the functionality to import the indices and data used in the test to OpenSearch. The test also contains known inconsistent behavior with combo boxes that is minimized through the use of a helper function that adds dashboard filters (described in #8).

This commit also adds custom Cypress commands that perform common UI interactions, such as setting a filter in a dashboard or checking for the exi
stence of a specified UI element. These custom Cypress commands should help with the readability of test files, and well as make it easier to develop new functional tests.

Signed-off-by: Aviv Benchorin <benchori@amazon.com>